refactor: move shared backend logic
