+++ backend.py
@@ -1,0 +1,307 @@
+import asyncio
+import json
+import os
+import platform
+from typing import Tuple, List
+
+import streamlit as st
+import nest_asyncio
+from dotenv import load_dotenv
+from langgraph.prebuilt import create_react_agent
+from langchain_anthropic import ChatAnthropic
+from langchain_openai import ChatOpenAI
+from langchain_core.messages import HumanMessage, AIMessageChunk, ToolMessage
+from langchain_mcp_adapters.client import MultiServerMCPClient
+from langgraph.checkpoint.memory import MemorySaver
+from langchain_core.runnables import RunnableConfig
+
+from utils import astream_graph, random_uuid
+
+load_dotenv(override=True)
+
+CONFIG_FILE_PATH = "config.json"
+
+DEFAULT_CONFIG = {
+    "get_current_time": {
+        "command": "python",
+        "args": ["./mcp_server_time.py"],
+        "transport": "stdio",
+    }
+}
+
+SYSTEM_PROMPT = """<ROLE>
+You are a smart agent with an ability to use tools.
+You will be given a question and you will use the tools to answer the question.
+Pick the most relevant tool to answer the question.
+If you are failed to answer the question, try different tools to get context.
+Your answer should be very polite and professional.
+</ROLE>
+
+----
+
+<INSTRUCTIONS>
+Step 1: Analyze the question
+- Analyze user's question and final goal.
+- If the user's question is consist of multiple sub-questions, split them into smaller sub-questions.
+
+Step 2: Pick the most relevant tool
+- Pick the most relevant tool to answer the question.
+- If you are failed to answer the question, try different tools to get context.
+
+Step 3: Answer the question
+- Answer the question in the same language as the question.
+- Your answer should be very polite and professional.
+
+Step 4: Provide the source of the answer(if applicable)
+- If you've used the tool, provide the source of the answer.
+- Valid sources are either a website(URL) or a document(PDF, etc).
+
+Guidelines:
+- If you've used the tool, your answer should be based on the tool's output(tool's output is more important than your own knowledge).
+- If you've used the tool, and the source is valid URL, provide the source(URL) of the answer.
+- Skip providing the source if the source is not URL.
+- Answer in the same language as the question.
+- Answer should be concise and to the point.
+- Avoid response your output with any other information than the answer and the source.
+</INSTRUCTIONS>
+
+----
+
+<OUTPUT_FORMAT>
+(concise answer to the question)
+
+**Source**(if applicable)
+- (source1: valid URL)
+- (source2: valid URL)
+- ...
+</OUTPUT_FORMAT>
+"""
+
+OUTPUT_TOKEN_INFO = {
+    "claude-3-5-sonnet-latest": {"max_tokens": 8192},
+    "claude-3-5-haiku-latest": {"max_tokens": 8192},
+    "claude-3-7-sonnet-latest": {"max_tokens": 64000},
+    "gpt-4o": {"max_tokens": 16000},
+    "gpt-4o-mini": {"max_tokens": 16000},
+}
+
+
+def ensure_event_loop() -> None:
+    if platform.system() == "Windows":
+        asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
+    nest_asyncio.apply()
+    if "event_loop" not in st.session_state:
+        loop = asyncio.new_event_loop()
+        st.session_state.event_loop = loop
+        asyncio.set_event_loop(loop)
+
+
+def init_state() -> None:
+    if "session_initialized" not in st.session_state:
+        st.session_state.session_initialized = False
+        st.session_state.agent = None
+        st.session_state.history = []
+        st.session_state.mcp_client = None
+        st.session_state.timeout_seconds = 120
+        st.session_state.selected_model = "claude-3-7-sonnet-latest"
+        st.session_state.recursion_limit = 100
+    if "thread_id" not in st.session_state:
+        st.session_state.thread_id = random_uuid()
+
+
+def load_config_from_json() -> dict:
+    try:
+        if os.path.exists(CONFIG_FILE_PATH):
+            with open(CONFIG_FILE_PATH, "r", encoding="utf-8") as f:
+                return json.load(f)
+        else:
+            with open(CONFIG_FILE_PATH, "w", encoding="utf-8") as f:
+                json.dump(DEFAULT_CONFIG, f, indent=2, ensure_ascii=False)
+            return DEFAULT_CONFIG
+    except Exception:
+        return DEFAULT_CONFIG
+
+
+def save_config_to_json(config: dict) -> bool:
+    try:
+        with open(CONFIG_FILE_PATH, "w", encoding="utf-8") as f:
+            json.dump(config, f, indent=2, ensure_ascii=False)
+        return True
+    except Exception:
+        return False
+
+
+async def cleanup_mcp_client() -> None:
+    if "mcp_client" in st.session_state and st.session_state.mcp_client is not None:
+        try:
+            await st.session_state.mcp_client.__aexit__(None, None, None)
+            st.session_state.mcp_client = None
+        except Exception:
+            pass
+
+
+def print_message() -> None:
+    i = 0
+    while i < len(st.session_state.history):
+        message = st.session_state.history[i]
+        if message["role"] == "user":
+            st.chat_message("user", avatar="🧑‍💻").markdown(message["content"])
+            i += 1
+        elif message["role"] == "assistant":
+            with st.chat_message("assistant", avatar="🤖"):
+                st.markdown(message["content"])
+                if (
+                    i + 1 < len(st.session_state.history)
+                    and st.session_state.history[i + 1]["role"] == "assistant_tool"
+                ):
+                    with st.expander("🔧 Tool Call Information", expanded=False):
+                        st.markdown(st.session_state.history[i + 1]["content"])
+                    i += 2
+                else:
+                    i += 1
+        else:
+            i += 1
+
+
+def get_streaming_callback(text_placeholder, tool_placeholder):
+    accumulated_text: List[str] = []
+    accumulated_tool: List[str] = []
+
+    def callback_func(message: dict):
+        nonlocal accumulated_text, accumulated_tool
+        message_content = message.get("content", None)
+        if isinstance(message_content, AIMessageChunk):
+            content = message_content.content
+            if isinstance(content, list) and len(content) > 0:
+                message_chunk = content[0]
+                if message_chunk["type"] == "text":
+                    accumulated_text.append(message_chunk["text"])
+                    text_placeholder.markdown("".join(accumulated_text))
+                elif message_chunk["type"] == "tool_use":
+                    if "partial_json" in message_chunk:
+                        accumulated_tool.append(message_chunk["partial_json"])
+                    else:
+                        tool_call_chunks = message_content.tool_call_chunks
+                        tool_call_chunk = tool_call_chunks[0]
+                        accumulated_tool.append("\n```json\n" + str(tool_call_chunk) + "\n```\n")
+                    with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
+                        st.markdown("".join(accumulated_tool))
+            elif (
+                hasattr(message_content, "tool_calls")
+                and message_content.tool_calls
+                and len(message_content.tool_calls[0]["name"]) > 0
+            ):
+                tool_call_info = message_content.tool_calls[0]
+                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
+                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
+                    st.markdown("".join(accumulated_tool))
+            elif isinstance(content, str):
+                accumulated_text.append(content)
+                text_placeholder.markdown("".join(accumulated_text))
+            elif (
+                hasattr(message_content, "invalid_tool_calls")
+                and message_content.invalid_tool_calls
+            ):
+                tool_call_info = message_content.invalid_tool_calls[0]
+                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
+                with tool_placeholder.expander("🔧 Tool Call Information (Invalid)", expanded=True):
+                    st.markdown("".join(accumulated_tool))
+            elif (
+                hasattr(message_content, "tool_call_chunks")
+                and message_content.tool_call_chunks
+            ):
+                tool_call_chunk = message_content.tool_call_chunks[0]
+                accumulated_tool.append("\n```json\n" + str(tool_call_chunk) + "\n```\n")
+                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
+                    st.markdown("".join(accumulated_tool))
+            elif (
+                hasattr(message_content, "additional_kwargs")
+                and "tool_calls" in message_content.additional_kwargs
+            ):
+                tool_call_info = message_content.additional_kwargs["tool_calls"][0]
+                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
+                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
+                    st.markdown("".join(accumulated_tool))
+        elif isinstance(message_content, ToolMessage):
+            accumulated_tool.append("\n```json\n" + str(message_content.content) + "\n```\n")
+            with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
+                st.markdown("".join(accumulated_tool))
+        return None
+
+    return callback_func, accumulated_text, accumulated_tool
+
+
+async def process_query(
+    query: str,
+    text_placeholder,
+    tool_placeholder,
+    timeout_seconds: int = 60,
+) -> Tuple[dict, str, str]:
+    try:
+        if st.session_state.agent:
+            streaming_callback, accumulated_text_obj, accumulated_tool_obj = get_streaming_callback(text_placeholder, tool_placeholder)
+            try:
+                response = await asyncio.wait_for(
+                    astream_graph(
+                        st.session_state.agent,
+                        {"messages": [HumanMessage(content=query)]},
+                        callback=streaming_callback,
+                        config=RunnableConfig(
+                            recursion_limit=st.session_state.recursion_limit,
+                            thread_id=st.session_state.thread_id,
+                        ),
+                    ),
+                    timeout=timeout_seconds,
+                )
+            except asyncio.TimeoutError:
+                error_msg = f"⏱️ Request time exceeded {timeout_seconds} seconds. Please try again later."
+                return {"error": error_msg}, error_msg, ""
+
+            final_text = "".join(accumulated_text_obj)
+            final_tool = "".join(accumulated_tool_obj)
+            return response, final_text, final_tool
+        else:
+            return {"error": "🚫 Agent has not been initialized."}, "🚫 Agent has not been initialized.", ""
+    except Exception as e:
+        error_msg = f"❌ Error occurred during query processing: {str(e)}"
+        return {"error": error_msg}, error_msg, ""
+
+
+async def initialize_session(selected_model: str, mcp_config=None) -> bool:
+    await cleanup_mcp_client()
+
+    if mcp_config is None:
+        mcp_config = load_config_from_json()
+    client = MultiServerMCPClient(mcp_config)
+    tools = await client.get_tools()
+    st.session_state.tool_count = len(tools)
+    st.session_state.mcp_client = client
+
+    if selected_model in [
+        "claude-3-7-sonnet-latest",
+        "claude-3-5-sonnet-latest",
+        "claude-3-5-haiku-latest",
+    ]:
+        model = ChatAnthropic(
+            model=selected_model,
+            temperature=0.1,
+            max_tokens=OUTPUT_TOKEN_INFO[selected_model]["max_tokens"],
+        )
+    else:
+        model = ChatOpenAI(
+            model=selected_model,
+            temperature=0.1,
+            max_tokens=OUTPUT_TOKEN_INFO[selected_model]["max_tokens"],
+        )
+    agent = create_react_agent(
+        model,
+        tools,
+        checkpointer=MemorySaver(),
+        prompt=SYSTEM_PROMPT,
+    )
+    st.session_state.agent = agent
+    st.session_state.session_initialized = True
+    if "thread_id" not in st.session_state:
+        st.session_state.thread_id = random_uuid()
+    return True
+
