+++ user_app.py
@@ -1,263 +1,22 @@
-import streamlit as st
-import asyncio
-import nest_asyncio
-import json
 import os
-import platform
+import streamlit as st
+
 from input_filter import InputFilter
+import backend
 
-if platform.system() == "Windows":
-    asyncio.set_event_loop_policy(asyncio.WindowsProactorEventLoopPolicy())
-
-nest_asyncio.apply()
-
-if "event_loop" not in st.session_state:
-    loop = asyncio.new_event_loop()
-    st.session_state.event_loop = loop
-    asyncio.set_event_loop(loop)
-
-from langgraph.prebuilt import create_react_agent
-from langchain_anthropic import ChatAnthropic
-from langchain_openai import ChatOpenAI
-from langchain_core.messages import HumanMessage
-from dotenv import load_dotenv
-from langchain_mcp_adapters.client import MultiServerMCPClient
-from utils import astream_graph, random_uuid
-from langgraph.checkpoint.memory import MemorySaver
-from langchain_core.messages.ai import AIMessageChunk
-from langchain_core.messages.tool import ToolMessage
-from langchain_core.runnables import RunnableConfig
-
-load_dotenv(override=True)
-
-CONFIG_FILE_PATH = "config.json"
-
-
-def load_config_from_json():
-    default_config = {
-        "get_current_time": {
-            "command": "python",
-            "args": ["./mcp_server_time.py"],
-            "transport": "stdio",
-        }
-    }
-    try:
-        if os.path.exists(CONFIG_FILE_PATH):
-            with open(CONFIG_FILE_PATH, "r", encoding="utf-8") as f:
-                return json.load(f)
-        else:
-            with open(CONFIG_FILE_PATH, "w", encoding="utf-8") as f:
-                json.dump(default_config, f, indent=2, ensure_ascii=False)
-            return default_config
-    except Exception:
-        return default_config
-
-
-SYSTEM_PROMPT = """<ROLE>
-You are a smart agent with an ability to use tools.
-You will be given a question and you will use the tools to answer the question.
-Pick the most relevant tool to answer the question.
-If you are failed to answer the question, try different tools to get context.
-Your answer should be very polite and professional.
-</ROLE>
-"""
-
-OUTPUT_TOKEN_INFO = {
-    "claude-3-5-sonnet-latest": {"max_tokens": 8192},
-    "claude-3-5-haiku-latest": {"max_tokens": 8192},
-    "claude-3-7-sonnet-latest": {"max_tokens": 64000},
-    "gpt-4o": {"max_tokens": 16000},
-    "gpt-4o-mini": {"max_tokens": 16000},
-}
-
-if "session_initialized" not in st.session_state:
-    st.session_state.session_initialized = False
-    st.session_state.agent = None
-    st.session_state.history = []
-    st.session_state.mcp_client = None
-    st.session_state.timeout_seconds = 120
-    st.session_state.selected_model = "claude-3-7-sonnet-latest"
-    st.session_state.recursion_limit = 100
-
-if "thread_id" not in st.session_state:
-    st.session_state.thread_id = random_uuid()
-
-
-async def cleanup_mcp_client():
-    if "mcp_client" in st.session_state and st.session_state.mcp_client is not None:
-        try:
-            await st.session_state.mcp_client.__aexit__(None, None, None)
-            st.session_state.mcp_client = None
-        except Exception:
-            pass
-
-
-def print_message():
-    i = 0
-    while i < len(st.session_state.history):
-        message = st.session_state.history[i]
-        if message["role"] == "user":
-            st.chat_message("user", avatar="🧑‍💻").markdown(message["content"])
-            i += 1
-        elif message["role"] == "assistant":
-            with st.chat_message("assistant", avatar="🤖"):
-                st.markdown(message["content"])
-                if (
-                    i + 1 < len(st.session_state.history)
-                    and st.session_state.history[i + 1]["role"] == "assistant_tool"
-                ):
-                    with st.expander("🔧 Tool Call Information", expanded=False):
-                        st.markdown(st.session_state.history[i + 1]["content"])
-                    i += 2
-                else:
-                    i += 1
-        else:
-            i += 1
-
-
-def get_streaming_callback(text_placeholder, tool_placeholder):
-    accumulated_text = []
-    accumulated_tool = []
-
-    def callback_func(message: dict):
-        nonlocal accumulated_text, accumulated_tool
-        message_content = message.get("content", None)
-        if isinstance(message_content, AIMessageChunk):
-            content = message_content.content
-            if isinstance(content, list) and len(content) > 0:
-                message_chunk = content[0]
-                if message_chunk["type"] == "text":
-                    accumulated_text.append(message_chunk["text"])
-                    text_placeholder.markdown("".join(accumulated_text))
-                elif message_chunk["type"] == "tool_use":
-                    if "partial_json" in message_chunk:
-                        accumulated_tool.append(message_chunk["partial_json"])
-                    else:
-                        tool_call_chunks = message_content.tool_call_chunks
-                        tool_call_chunk = tool_call_chunks[0]
-                        accumulated_tool.append("\n```json\n" + str(tool_call_chunk) + "\n```\n")
-                    with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
-                        st.markdown("".join(accumulated_tool))
-            elif (
-                hasattr(message_content, "tool_calls")
-                and message_content.tool_calls
-                and len(message_content.tool_calls[0]["name"]) > 0
-            ):
-                tool_call_info = message_content.tool_calls[0]
-                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
-                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
-                    st.markdown("".join(accumulated_tool))
-            elif isinstance(content, str):
-                accumulated_text.append(content)
-                text_placeholder.markdown("".join(accumulated_text))
-            elif (
-                hasattr(message_content, "invalid_tool_calls")
-                and message_content.invalid_tool_calls
-            ):
-                tool_call_info = message_content.invalid_tool_calls[0]
-                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
-                with tool_placeholder.expander("🔧 Tool Call Information (Invalid)", expanded=True):
-                    st.markdown("".join(accumulated_tool))
-            elif (
-                hasattr(message_content, "tool_call_chunks")
-                and message_content.tool_call_chunks
-            ):
-                tool_call_chunk = message_content.tool_call_chunks[0]
-                accumulated_tool.append("\n```json\n" + str(tool_call_chunk) + "\n```\n")
-                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
-                    st.markdown("".join(accumulated_tool))
-            elif (
-                hasattr(message_content, "additional_kwargs")
-                and "tool_calls" in message_content.additional_kwargs
-            ):
-                tool_call_info = message_content.additional_kwargs["tool_calls"][0]
-                accumulated_tool.append("\n```json\n" + str(tool_call_info) + "\n```\n")
-                with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
-                    st.markdown("".join(accumulated_tool))
-        elif isinstance(message_content, ToolMessage):
-            accumulated_tool.append("\n```json\n" + str(message_content.content) + "\n```\n")
-            with tool_placeholder.expander("🔧 Tool Call Information", expanded=True):
-                st.markdown("".join(accumulated_tool))
-        return None
-
-    return callback_func, accumulated_text, accumulated_tool
-
-
-async def process_query(query, text_placeholder, tool_placeholder, timeout_seconds=60):
-    try:
-        if st.session_state.agent:
-            streaming_callback, accumulated_text_obj, accumulated_tool_obj = get_streaming_callback(text_placeholder, tool_placeholder)
-            try:
-                response = await asyncio.wait_for(
-                    astream_graph(
-                        st.session_state.agent,
-                        {"messages": [HumanMessage(content=query)]},
-                        callback=streaming_callback,
-                        config=RunnableConfig(
-                            recursion_limit=st.session_state.recursion_limit,
-                            thread_id=st.session_state.thread_id,
-                        ),
-                    ),
-                    timeout=timeout_seconds,
-                )
-            except asyncio.TimeoutError:
-                error_msg = f"⏱️ Request time exceeded {timeout_seconds} seconds. Please try again later."
-                return {"error": error_msg}, error_msg, ""
-
-            final_text = "".join(accumulated_text_obj)
-            final_tool = "".join(accumulated_tool_obj)
-            return response, final_text, final_tool
-        else:
-            return {"error": "🚫 Agent has not been initialized."}, "🚫 Agent has not been initialized.", ""
-    except Exception as e:
-        error_msg = f"❌ Error occurred during query processing: {str(e)}"
-        return {"error": error_msg}, error_msg, ""
-
-
-async def initialize_session(mcp_config=None):
-    with st.spinner("🔄 Connecting to MCP server..."):
-        await cleanup_mcp_client()
-        if mcp_config is None:
-            mcp_config = load_config_from_json()
-        client = MultiServerMCPClient(mcp_config)
-        tools = await client.get_tools()
-        st.session_state.mcp_client = client
-        selected_model = st.session_state.selected_model
-        if selected_model in [
-            "claude-3-7-sonnet-latest",
-            "claude-3-5-sonnet-latest",
-            "claude-3-5-haiku-latest",
-        ]:
-            model = ChatAnthropic(
-                model=selected_model,
-                temperature=0.1,
-                max_tokens=OUTPUT_TOKEN_INFO[selected_model]["max_tokens"],
-            )
-        else:
-            model = ChatOpenAI(
-                model=selected_model,
-                temperature=0.1,
-                max_tokens=OUTPUT_TOKEN_INFO[selected_model]["max_tokens"],
-            )
-        agent = create_react_agent(
-            model,
-            tools,
-            checkpointer=MemorySaver(),
-            prompt=SYSTEM_PROMPT,
-        )
-        st.session_state.agent = agent
-        st.session_state.session_initialized = True
-        return True
-
+backend.ensure_event_loop()
+backend.init_state()
 
 st.set_page_config(page_title="Chat", page_icon="🧠", layout="wide")
 
 st.title("💬 MCP Tool Chat")
 
 if not st.session_state.session_initialized:
-    st.session_state.event_loop.run_until_complete(initialize_session())
+    st.session_state.event_loop.run_until_complete(
+        backend.initialize_session(st.session_state.selected_model)
+    )
 
-print_message()
+backend.print_message()
 
 user_query = st.chat_input("💬 Enter your question")
 if user_query:
@@ -272,7 +31,7 @@ if user_query:
         tool_placeholder = st.empty()
         text_placeholder = st.empty()
         resp, final_text, final_tool = st.session_state.event_loop.run_until_complete(
-            process_query(
+            backend.process_query(
                 user_query,
                 text_placeholder,
                 tool_placeholder,
@@ -287,3 +46,4 @@ if user_query:
         if final_tool.strip():
             st.session_state.history.append({"role": "assistant_tool", "content": final_tool})
         st.rerun()
+
